requirements.txt
requests
tqdm
httpx
aiofiles
python-dotenv
fastapi
uvicorn
jinja2
websockets
